[ClangRepl] support code completion at the repl
This patch enabled code completion for ClangREPL. The feature was built upon two
existing Clang components: a list completer for `LineEditor` and `SemaCodeCompletion`.

The former serves the main entry point to trigger the latter. Because a
completion point for a compiler instance needs to be unchanged once it is set,
an incremental compiler instance for code completion is created. In addition to
completion points, it differs from a regular incremental compiler in the following ways:

1. It does not execute input.

2. To obtain declarations or bindings from previous input in the same REPL
session, it carries over AST context source from the main compiler, i.e. the one
used in the interpreter.

3. It contains a `ReplCompletionConsumer`, a subclass of
`CodeCompletionConsumer`. The consumer communicates completion results from
`SemaCodeCompletion` back to the list completer for the REPL.

In addition, `PCC_TopLevelStmtDecl and `CCC_ReplTopLevel` top levels were added so
that `SemaCodeCompletion` can treat top level statements at the REPL like
expression statements. For example,

```
clang-repl> int foo = 42;
clang-repl> f<tab>
```

From a parser's persective, the cursor is at a top level. If we used code
completion without any changes, `PCC_Namespace` would be supplied to
`Sema::CodeCompleteOrdinaryName`, and thus the completion results would not
include `foo`.

Currently, the way we use `PCC_TopLevelStmtDecl` and `CCC_ReplTopLevel` is no
different from the way we use `PCC_Statement` and `CCC_Statement` respectively.

Differential Revision: https://reviews.llvm.org/D154382